Guards against invalid input in buy form.
Removed invalid import (Observable.of to of) in spec file.
Added buyInputValidator to prevent users from entering non digit values
(deletes invalid values in form automatically).
Added test for buyInputValidator.
Added testInteger method to prevent users from entering non digit values
(this is for an extra layer of check internally in the buy and sell methods).
Added tests for testInteger method.
Fixed bug where the stockDetails page is unreachable because the stocks
array was not populated before it was used in the check condition.
Added checks in buy and sell methods to prevent decimal numbers from being
entered.